EventListener demonstration
Spring 4.2 makes it very easy to declare an event listener. Just add the
`@EventListener` annotation on a method of a bean with the type of event
you are listening to.

Spring Data REST is actually sending various events related to
repositories actions. We add a listener for `AfterCreatedEvent` that only
prints the entity that was just created.

We can post the json representation of a book to `/data/books` to see
what happens.