Add basic book order workflow
Add a very basic endpoint that allow one to order a book. This creates
a `BookOrder` and verify the order before returning. We showcase the use
of transaction management for the processing of the order as well as how
one can use Spring Data REST entity link infrastructure to automatically
generate links to related content.

We also use the `ExceptionHandler` infrastructure to map our specific
exception to a well-defined status code and reason.