Additional event listener demonstration
We separate our ordering workflow in two separate steps that are designed
as events. Previously, one had to wrap such instance in application
events; this is no longer the case.

The first event listener returns the status. This will send a new event
behind the scenes that is handled by our second listener.

Of course, our workflow is still slow as events are handled in a
synchronous fashion by default. We could chang that or use a message
broker instead.