JmsListener demonstration
Update our `BookOrderService` to use a JMS broker instead of the event
infrastructure. For the sake of the demo, we are going to use an embedded
HornetQ broker.

We need first to add the hornetQ starter as well as the
`hornetq-jms-server` dependency that brings the server along. Once we've
done that we need to replace our `@EventListener` with the `@JmsListener`
counterpart. We also need to add a `@SendTo` annotation on the first
listener to specify on which destination the reply should be sent. We
also need to replace our event publisher by a messaging template for JMS.

Finally, there are two ways to define the queues to use on startup. We
could use the configuration support or we could define a `@Bean` of type
`JMSQueueConfiguration`. We chose the latter approach.