Basic cache demonstration
If we enable `org.hibernate.SQL` we will see that we hit the database
every time we need to retrieve a book by id. There are multiple places
where we could add caching but for the sake of the demo we're going to
add some basic cache instructions on the repository itself.

First, let's add the cache starter and add `@EnableCaching` on our Spring
Boot application. With Spring Boot 1.3, this is all that's needed to
configure the caching infrastructure.

Then we add a `@Cacheable` on `findOne` and a `@CacheEvict` on
`delete`. Again, more cache semantics should be added to make sure you
don't return staled data.

If we run our app again, we can see that the SQL is only issued once per
book. We can also hit the `/metrics` endpoint to show the cache metrics
that Spring Boot exposes out-of-the-box. Since we haven't specified
anything, Spring Boot is using a basic concurrent hash map
implementation.