JSR-107 demonstration
Spring Boot allows to integrate to a JSR-107 cache manager if it is
available on the classpath. We add ehcache3 and the cache-api to our
project which is all that's needed to start using JSR-107.

We can customize the `javax.cache.CacheManager` easily by registering a
bean of type `JCacheManagerCustomizer`. We can create our `books cache
with a custom configuration.

If we run our app again, we can see that the `/metrics` endpoint exposes
cache hit and miss ratio now.

Finally, we can update our `BookRepository` to use the standard JSR-107
annotations. This last step is by no mean necessary: you can also use a
JSR-107 cache manager with the regular Spring cache annotations.