Generic injection demonstration
To demonstrate how Spring Framework can resolve generic information to
inject the right component, we can create a common interface for all our
repositories that exposes the type of the entity. We also should add the
`@NoRepositoryBean` on that interface as it is not an actual repository.

Then we can update `BookOrderService` to inject that base interface
rather than the actual repository type.